pyproject.toml
[project]
name = "lex-db"
version = "0.1.0"
description = "A wrapper around a sqlite database for encyclopedia articles with vector and full-text search capabilities"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "sqlite-vec>=0.1.6",
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "ruff>=0.11.9",
    "mypy>=1.15.0",
    "pytest>=8.3.5",
    "pytest-cov>=6.1.1",
    "sentence-transformers>=4.1.0",
    "openai>=1.79.0",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.7.0",
    "pytest>=7.4.0",
]
